refactor(frontend): decompose ProductivityScreen God Class
Debt Identified: The `frontend/app/(main)/productivity.tsx` was an 850+ line "God Class" handling complex state, filtering algorithms, daily plan generation logic, and rendering deeply nested UI components. This violated the Single Responsibility Principle and made the widget difficult to maintain.

Refactored Code:
- Extracted UI sections into dedicated components: `ProductivityHeader`, `ProductivityTabs`, `ProductivityEmptyState`, and `ProductivityTodayPlan` under `frontend/src/components/productivity/`.
- Decoupled business logic into `useProductivityScreen` custom hook under `frontend/src/hooks/`.
- Edited `productivity.tsx` to strictly compose the view without leaking logic.

Structural Note: The change strictly follows SOLID principles. The main screen acts purely as a coordinator (Single Responsibility), view logic is decoupled into the hook (Dependency Inversion), and components are modularized to under 150 lines, improving testability and code readability without altering behavior.
