+++ frontend/src/components/productivity/ProductivityTabs.tsx
@@ -1,0 +1,173 @@
+import React from 'react';
+import { View, StyleSheet, Pressable } from 'react-native';
+import { useTranslation } from 'react-i18next';
+import { AppText } from '../AppText';
+import { theme } from '../../core/theme';
+import { DESIGN_TOKENS } from '@/core/design/tokens';
+
+const S = theme.spacing;
+const R = theme.borderRadius;
+
+const T = {
+  background: { light: DESIGN_TOKENS.colors.pageBg },
+  surface: { light: DESIGN_TOKENS.colors.surface, highLight: DESIGN_TOKENS.colors.surfaceSoft },
+  border: { light: DESIGN_TOKENS.colors.border },
+  onSurface: {
+    light: DESIGN_TOKENS.colors.text,
+    mutedLight: DESIGN_TOKENS.colors.muted,
+    disabledLight: DESIGN_TOKENS.colors.faint,
+  },
+  primary: {
+    DEFAULT: DESIGN_TOKENS.colors.primary,
+    surfaceLight: DESIGN_TOKENS.colors.primarySoft,
+  },
+  white: '#FFFFFF',
+  transparent: 'transparent',
+};
+
+export type Tab = 'today' | 'all' | 'notes' | 'tasks';
+export type TaskPriority = 'all' | 'overdue' | 'today' | 'upcoming' | 'no_due';
+
+interface Props {
+  activeTab: Tab;
+  setActiveTab: (tab: Tab) => void;
+  taskPriority: TaskPriority;
+  setTaskPriority: (priority: TaskPriority) => void;
+  taskPriorityCounts: Record<TaskPriority, number>;
+}
+
+export const ProductivityTabs = React.memo(({
+  activeTab,
+  setActiveTab,
+  taskPriority,
+  setTaskPriority,
+  taskPriorityCounts,
+}: Props) => {
+  const { t } = useTranslation();
+
+  return (
+    <View>
+      <View style={styles.tabsContainer}>
+        {(['today', 'all', 'notes', 'tasks'] as const).map((tab) => (
+          <Pressable
+            key={tab}
+            onPress={() => setActiveTab(tab)}
+            style={({ pressed }) => [
+              styles.tabButton,
+              activeTab === tab && styles.tabButtonActive,
+              pressed && activeTab !== tab && { opacity: 0.6 },
+            ]}
+          >
+            <AppText style={[styles.tabText, activeTab === tab && styles.tabTextActive]}>
+              {tab === 'today'
+                ? t('productivity.today')
+                : tab === 'all'
+                  ? t('productivity.all') || 'All'
+                  : tab === 'notes'
+                    ? t('productivity.notes') || 'Notes'
+                    : t('productivity.tasks') || 'Tasks'}
+            </AppText>
+          </Pressable>
+        ))}
+      </View>
+
+      {(activeTab === 'tasks' || activeTab === 'today') && (
+        <View
+          style={{
+            flexDirection: 'row',
+            flexWrap: 'wrap',
+            marginHorizontal: S.xl,
+            marginBottom: S.sm,
+          }}
+        >
+          {(
+            [
+              { key: 'all', label: t('productivity.priority.all', { count: taskPriorityCounts.all }) },
+              {
+                key: 'overdue',
+                label: t('productivity.priority.overdue', { count: taskPriorityCounts.overdue }),
+              },
+              {
+                key: 'today',
+                label: t('productivity.priority.today', { count: taskPriorityCounts.today }),
+              },
+              {
+                key: 'upcoming',
+                label: t('productivity.priority.upcoming', { count: taskPriorityCounts.upcoming }),
+              },
+              {
+                key: 'no_due',
+                label: t('productivity.priority.no_due', { count: taskPriorityCounts.no_due }),
+              },
+            ] as const
+          ).map((option) => (
+            <Pressable
+              key={option.key}
+              onPress={() => setTaskPriority(option.key)}
+              style={({ pressed }) => [
+                {
+                  borderRadius: 12,
+                  borderWidth: 1,
+                  borderColor: taskPriority === option.key ? T.primary.DEFAULT : T.border.light,
+                  backgroundColor:
+                    taskPriority === option.key ? T.primary.surfaceLight : T.surface.light,
+                  paddingHorizontal: 10,
+                  paddingVertical: 6,
+                  marginRight: 8,
+                  marginBottom: 8,
+                },
+                pressed && { opacity: 0.8 },
+              ]}
+            >
+              <AppText
+                variant="caption"
+                style={{
+                  color: taskPriority === option.key ? T.primary.DEFAULT : T.onSurface.mutedLight,
+                  fontWeight: '700',
+                }}
+              >
+                {option.label}
+              </AppText>
+            </Pressable>
+          ))}
+        </View>
+      )}
+    </View>
+  );
+});
+
+ProductivityTabs.displayName = 'ProductivityTabs';
+
+const styles = StyleSheet.create({
+  tabsContainer: {
+    flexDirection: 'row',
+    backgroundColor: T.surface.highLight,
+    borderRadius: R.xl,
+    padding: S.xs,
+    marginHorizontal: S.xl,
+    marginTop: S.lg,
+    marginBottom: S.md,
+    borderWidth: 1,
+    borderColor: T.border.light,
+  },
+  tabButton: {
+    flex: 1,
+    paddingVertical: S.sm,
+    alignItems: 'center',
+    borderRadius: R.lg,
+    backgroundColor: T.transparent,
+  },
+  tabButtonActive: {
+    backgroundColor: T.surface.light,
+    ...theme.elevation.sm,
+  },
+  tabText: {
+    fontSize: 14,
+    fontWeight: '500',
+    color: T.onSurface.mutedLight,
+  },
+  tabTextActive: {
+    fontWeight: '700',
+    color: T.onSurface.light,
+  },
+});
